Bug exploration

+++ ProjectFlightSchool/ProjectFlightSchool/Structs.h
@@ -49,7 +49,7 @@ struct EvProjectileFired
 struct EvMeleeHit
 {
 	unsigned int			ID;
-	unsigned int			damage;
+	float					damage;
 	float					knockBack;
 	DirectX::XMFLOAT3		direction;
 };
